upgraded vico to 2.0.0-alpha.24

+++ app/src/main/kotlin/edu/stanford/bdh/engagehf/health/HealthUiStateMapper.kt
@@ -52,7 +52,8 @@ class HealthUiStateMapper @Inject constructor(
     ): HealthUiData {
         val filteredRecords: List<EngageRecord> =
             filterRecordsByTimeRange(records, selectedTimeRange)
-        val pairs: List<Pair<Float, Float>> = groupAndMapRecords(filteredRecords, selectedTimeRange)
+        val pairs: List<Pair<Double, Double>> =
+            groupAndMapRecords(filteredRecords, selectedTimeRange)
         val title: String = when (records.first()) {
             is EngageRecord.HeartRate -> "Heart Rate"
             is EngageRecord.BloodPressure -> "Blood Pressure"
@@ -62,7 +63,7 @@ class HealthUiStateMapper @Inject constructor(
 
         val chartDataList = mutableListOf(chartData)
         if (records.any { it is EngageRecord.BloodPressure }) {
-            val diastolicPairs: List<Pair<Float, Float>> =
+            val diastolicPairs: List<Pair<Double, Double>> =
                 groupAndMapRecords(filteredRecords, selectedTimeRange, true)
             val diastolicChartData = createAggregatedHealthData(title, diastolicPairs)
             chartDataList.add(diastolicChartData)
@@ -98,22 +99,22 @@ class HealthUiStateMapper @Inject constructor(
         records: List<EngageRecord>,
         selectedTimeRange: TimeRange,
         diastolic: Boolean = false,
-    ): List<Pair<Float, Float>> {
+    ): List<Pair<Double, Double>> {
         val groupedRecords = groupRecordsByTimeRange(records, selectedTimeRange)
         return groupedRecords.values.map { entries ->
             val averageValue = entries.map { getValue(it, diastolic).value }.average()
             Pair(
-                averageValue.toFloat().roundToDecimalPlaces(2),
+                averageValue.roundToDecimalPlaces(2),
                 mapXValue(selectedTimeRange, entries.first().zonedDateTime).roundToDecimalPlaces(2)
             )
         }
     }
 
-    private fun mapXValue(selectedTimeRange: TimeRange, zonedDateTime: ZonedDateTime): Float {
+    private fun mapXValue(selectedTimeRange: TimeRange, zonedDateTime: ZonedDateTime): Double {
         return when (selectedTimeRange) {
-            TimeRange.DAILY -> (zonedDateTime.year.toFloat() + (zonedDateTime.dayOfYear - 1) / 365f) * 10
-            TimeRange.WEEKLY -> (zonedDateTime.toEpochSecond() / (7 * 24 * 60 * 60)).toFloat()
-            TimeRange.MONTHLY -> zonedDateTime.year.toFloat() + (zonedDateTime.monthValue - 1) / 12f
+            TimeRange.DAILY -> (zonedDateTime.year.toDouble() + (zonedDateTime.dayOfYear - 1) / 365.0) * 10
+            TimeRange.WEEKLY -> (zonedDateTime.toEpochSecond() / (7 * 24 * 60 * 60)).toDouble()
+            TimeRange.MONTHLY -> zonedDateTime.year.toDouble() + (zonedDateTime.monthValue - 1) / 12f
         }
     }
 
@@ -207,7 +208,7 @@ class HealthUiStateMapper @Inject constructor(
 
     private fun createAggregatedHealthData(
         title: String,
-        pairs: List<Pair<Float, Float>>,
+        pairs: List<Pair<Double, Double>>,
     ): AggregatedHealthData {
         return AggregatedHealthData(
             yValues = pairs.map { it.first },
@@ -242,7 +243,7 @@ class HealthUiStateMapper @Inject constructor(
 
             val tableEntryData = TableEntryData(
                 id = currentRecord.clientRecordId,
-                value = currentRecordValue.value.toFloat(),
+                value = currentRecordValue.value,
                 secondValue = if (currentRecord is EngageRecord.BloodPressure) {
                     getValue(currentRecord, true).value.toFloat()
                 } else {
@@ -250,7 +251,7 @@ class HealthUiStateMapper @Inject constructor(
                 },
                 date = currentRecord.zonedDateTime,
                 formattedDate = currentRecord.zonedDateTime.format(monthDayTimeFormatter),
-                trend = trend.toFloat(),
+                trend = trend,
                 formattedTrend = formattedTrend,
                 formattedValues = if (currentRecord is EngageRecord.BloodPressure) {
                     getBloodPressureFormatRecord(currentRecord.record)
@@ -268,7 +269,7 @@ class HealthUiStateMapper @Inject constructor(
 
     private fun getAverageData(tableData: List<TableEntryData>) =
         AverageHealthData(
-            value = tableData.mapNotNull { it.value }.average().toFloat(),
+            value = tableData.mapNotNull { it.value }.average(),
             formattedValue = "Average " + formatValue(tableData.mapNotNull { it.value }.average()),
         )
 
@@ -329,7 +330,7 @@ class HealthUiStateMapper @Inject constructor(
 
     private fun getDefaultLocale() = localeProvider.getDefaultLocale()
 
-    private fun valueFormatter(value: Float, timeRange: TimeRange): String {
+    private fun valueFormatter(value: Double, timeRange: TimeRange): String {
         val date = when (timeRange) {
             TimeRange.DAILY -> {
                 val actualValue = value * 10
